Update components  (#9)
* update about compt btn links

* add title card compt, import to LMap
